Move `node-internal:inspect` into `node-internal:util`
