+++ src/workerd/api/node/util.c++
@@ -127,4 +127,59 @@ kj::String MIMEType::toString() {
   return inner.toString();
 }
 
+jsg::JsArray UtilModule::getOwnNonIndexProperties(jsg::Lock& js, jsg::JsObject value,
+                                                       int filter) {
+  auto propertyFilter = static_cast<jsg::PropertyFilter>(filter);
+  return value.getPropertyNames(js, jsg::KeyCollectionFilter::OWN_ONLY, propertyFilter,
+                                  jsg::IndexFilter::SKIP_INDICES);
+}
+
+jsg::Optional<UtilModule::PromiseDetails> UtilModule::getPromiseDetails(jsg::JsValue value) {
+  auto promise = KJ_UNWRAP_OR_RETURN(value.tryCast<jsg::JsPromise>(), kj::none);
+  auto state = promise.state();
+  if (state != jsg::PromiseState::PENDING) {
+    auto result = promise.result();
+    return PromiseDetails { .state = state, .result = result };
+  } else {
+    return PromiseDetails { .state = state, .result = kj::none };
+  }
+}
+
+jsg::Optional<UtilModule::ProxyDetails> UtilModule::getProxyDetails(jsg::JsValue value) {
+  auto proxy = KJ_UNWRAP_OR_RETURN(value.tryCast<jsg::JsProxy>(), kj::none);
+  auto target = proxy.target();
+  auto handler = proxy.handler();
+  return ProxyDetails { .target = target, .handler = handler };
+}
+
+jsg::Optional<UtilModule::PreviewedEntries> UtilModule::previewEntries(jsg::JsValue value) {
+  auto object = KJ_UNWRAP_OR_RETURN(value.tryCast<jsg::JsObject>(), kj::none);
+  bool isKeyValue;
+  auto entries = object.previewEntries(&isKeyValue);
+  return PreviewedEntries { .entries = entries, .isKeyValue = isKeyValue };
+}
+
+jsg::JsString UtilModule::getConstructorName(jsg::Lock& js, jsg::JsObject value) {
+  return js.str(value.getConstructorName());
+}
+
+#define V(Type) bool UtilModule::is##Type(jsg::JsValue value) { return value.is##Type(); };
+  JS_UTIL_IS_TYPES(V)
+#undef V
+bool UtilModule::isAnyArrayBuffer(jsg::JsValue value) {
+  return value.isArrayBuffer() || value.isSharedArrayBuffer();
+}
+bool UtilModule::isBoxedPrimitive(jsg::JsValue value) {
+  return value.isNumberObject() ||
+         value.isStringObject() ||
+         value.isBooleanObject() ||
+         value.isBigIntObject() ||
+         value.isSymbolObject();
+}
+
+jsg::Name UtilModule::getResourceTypeInspect(jsg::Lock& js) {
+  return js.newApiSymbol("kResourceTypeInspect"_kj);
+}
+
+
 }  // namespace workerd::api::node
